+++ contextor/core/live_state/__init__.py
@@ -10,6 +10,7 @@ from .store import (
 from .ipc import CanonicalLiveServer, LiveEndpoint, LiveStateClient
 from .runtime import connect, connect_or_start
 from .watcher import DesktopLiveEventFeed, DesktopLiveWatcher
+from .hydration import HydratedRepositoryEngine, hydrate_repository_engine
 
 __all__ = [
     "CanonicalLiveServer",
@@ -18,8 +19,10 @@ __all__ = [
     "LiveStateMetadata",
     "DesktopLiveWatcher",
     "DesktopLiveEventFeed",
+    "HydratedRepositoryEngine",
     "connect",
     "connect_or_start",
+    "hydrate_repository_engine",
     "load_snapshot",
     "migrate_legacy_snapshot",
     "read_metadata",